+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 * Added support for exporting textures to a model.nut file on export.
 
+### Fixed
+* Fixed an issue where global textures like `10080000` would not import images.
+
 ## 0.4.0 - 2026-03-17
 ### Changed
 * Changed supported Blender version to 5.1 or later.